+++ fresh.gen.ts
@@ -19,7 +19,8 @@ import * as $layout_index from "./routes/layout/index.tsx";
 import * as $markdown_index from "./routes/markdown/index.tsx";
 import * as $panel_panel from "./routes/panel/panel.tsx";
 import * as $text_text from "./routes/text/text.tsx";
-import * as $AddChip from "./islands/AddChip.tsx";
+import * as $Autocomplete from "./islands/Autocomplete.tsx";
+import * as $KvAutocomplete from "./islands/KvAutocomplete.tsx";
 import { type Manifest } from "$fresh/server.ts";
 
 const manifest = {
@@ -43,7 +44,8 @@ const manifest = {
     "./routes/text/text.tsx": $text_text,
   },
   islands: {
-    "./islands/AddChip.tsx": $AddChip,
+    "./islands/Autocomplete.tsx": $Autocomplete,
+    "./islands/KvAutocomplete.tsx": $KvAutocomplete,
   },
   baseUrl: import.meta.url,
 } satisfies Manifest;